Refactor into per-type workflows:
The syndication sources and proxies, while involved with each other's
workflows, don't share the same workflow *state*.  IMO we shouldn't be abusing
a single workflow when the different types we apply it do don't really have
the same state.  Similarly, the syndication state is distinct from the move
state: a given source can be syndicated *and* moved, not syndicated and moved,
syndicated and not moved, or not syndicated and not moved.  So the syndication
and move states should also have their own workflows.

This change separates the source workflow from the proxy workflow and the
syndication workflows from the move workflows.  The accept/reject transitions
on the source object are handled by automatic transitions that are guarded by
checking if all proxies are accepted/rejected.  This should make it much
harder for sources and proxies to get out of sync in terms of state.